remove stringifyVtt and add vtt support to stringify
