+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "subtitle",
-  "version": "2.0.5",
+  "version": "3.0.0",
   "description": "Parse and manipulate SRT (SubRip)",
   "repository": {
     "type": "git",
