+++ src/index.ts
@@ -3,7 +3,6 @@ export * from './toSrtTime'
 export * from './toVttTime'
 export * from './parse'
 export * from './stringify'
-export * from './stringifyVtt'
 export * from './resync'
 export * from './parseTimestamps'
 export * from './types'
